Update MQUnifiedsensor.cpp
Fixed compile warnings in MQUnifiedsensor.cpp by correcting I2C requestFrom() function call

- Cast both parameters (address and size) in Wire.requestFrom() to uint8_t to ensure proper type matching with the Wire library function.
- Resolved type mismatch issue that caused compile warnings.

+++ src/libs/MQUnifiedsensor/src/MQUnifiedsensor.cpp
@@ -166,7 +166,7 @@ void MQUnifiedsensor::update()
     else
     {
         uint8_t temp[2];
-        Wire.requestFrom(addr, 2);
+        Wire.requestFrom((uint8_t)addr, (uint8_t)2);  // Fix: Cast both parameters to uint8_t
         temp[0] = Wire.read();
         temp[1] = Wire.read();
         _adc = (temp[0] | temp[1] << 8);
